@@ -3,10 +3,22 @@
   "version": "0.0.0",
   "main": "server.js",
   "scripts": {
-    "start": "nodemon server.js",
-    "migrate": "npx sequelize db:migrate",
-    "add-migration": "npx sequelize migration:generate --name",
-    "seed": "npx sequelize db:seed:undo:all && npx sequelize db:seed:all"
+    "start": "node --inspect=5858 -r ts-node/register ./src/server.ts",
+    "start:watch": "nodemon",
+    "build": "tsc"
+  },
+  "nodemonConfig": {
+    "ignore": [
+      "**/*.test.ts",
+      "**/*.spec.ts",
+      ".git",
+      "node_modules"
+    ],
+    "watch": [
+      "src"
+    ],
+    "exec": "npm start",
+    "ext": "ts"
   },
   "dependencies": {
     "cors": "^2.8.5",
@@ -24,9 +36,15 @@
     "uuid": "^3.3.3"
   },
   "devDependencies": {
-    "eslint": "^6.4.0",
+    "@types/cors": "^2.8.6",
+    "@types/express-fileupload": "^1.1.3",
+    "@types/mongoose": "^5.7.30",
+    "@types/node": "^14.0.20",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "nodemon": "^1.19.2"
+    "eslint-plugin-import": "^2.22.0",
+    "nodemon": "^1.19.4",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.6"
   }
 }
